compile

+++ db/javajs.cpp
@@ -3,7 +3,6 @@
 /**
 *    Copyright (C) 2008 10gen Inc.
 *  
-*    This program is free software: you can redistribute it and/or  modify
 *    it under the terms of the GNU Affero General Public License, version 3,
 *    as published by the Free Software Foundation.
 *  
@@ -154,9 +153,9 @@ JavaJSImpl::JavaJSImpl(const char *appserverPath){
   }
 
   if( res ) {
-	  problem() << "Couldn't create JVM res:" << res << " terminating" << endl;
-	  log() << "(try --nojni if you do not require that functionality)" << endl;
-	  exit(22);
+    problem() << "Couldn't create JVM res:" << (int) res << " terminating" << endl;
+    log() << "(try --nojni if you do not require that functionality)" << endl;
+    exit(22);
   }
   jassert( res == 0 );
   jassert( _jvm > 0 );
